Update ReviverError67
This is the current update for the error reviewed in our meeting today. As we can see, the error67 only appears when the user inputs different data, but this was resolved and fixed with the updated code.

If you have any questions, feel free to message Jime or me or you can edit the code freely if any solution arises.